Move uninstall inside commands.
