+++ hatch/commands/__init__.py
@@ -4,3 +4,4 @@ from hatch.commands.config import config
 from hatch.commands.new import new
 from hatch.commands.init import init
 from hatch.commands.install import install
+from hatch.commands.uninstall import uninstall
